+++ lib/Models/WebMapServiceCatalogItem.ts
@@ -30,11 +30,15 @@ import ExportableMixin from "../ModelMixins/ExportableMixin";
 import GetCapabilitiesMixin from "../ModelMixins/GetCapabilitiesMixin";
 import TimeFilterMixin from "../ModelMixins/TimeFilterMixin";
 import UrlMixin from "../ModelMixins/UrlMixin";
+import SelectableDimensions, {
+  SelectableDimension
+} from "../Models/SelectableDimensions";
 import { InfoSectionTraits } from "../Traits/CatalogMemberTraits";
 import DiscreteTimeTraits from "../Traits/DiscreteTimeTraits";
 import LegendTraits from "../Traits/LegendTraits";
 import { RectangleTraits } from "../Traits/MappableTraits";
 import WebMapServiceCatalogItemTraits, {
+  WebMapServiceAvailableLayerDimensionsTraits,
   WebMapServiceAvailableLayerStylesTraits
 } from "../Traits/WebMapServiceCatalogItemTraits";
 import { callWebCoverageService } from "./callWebCoverageService";
@@ -48,28 +52,13 @@ import proxyCatalogItemUrl from "./proxyCatalogItemUrl";
 import StratumFromTraits from "./StratumFromTraits";
 import WebMapServiceCapabilities, {
   CapabilitiesContactInformation,
+  CapabilitiesDimension,
   CapabilitiesLayer,
   CapabilitiesStyle,
   getRectangleFromLayer
 } from "./WebMapServiceCapabilities";
 
 const dateFormat = require("dateformat");
-
-interface LegendUrl {
-  url: string;
-  mimeType?: string;
-}
-
-interface WebMapServiceStyle {
-  name: string;
-  title: string;
-  abstract?: string;
-  legendUrl?: LegendUrl;
-}
-
-interface WebMapServiceStyles {
-  [layerName: string]: WebMapServiceStyle[];
-}
 
 class GetCapabilitiesStratum extends LoadableStratum(
   WebMapServiceCatalogItemTraits
@@ -194,6 +183,50 @@ class GetCapabilitiesStratum extends LoadableStratum(
   }
 
   @computed
+  get availableDimensions(): StratumFromTraits<
+    WebMapServiceAvailableLayerDimensionsTraits
+  >[] {
+    const result: StratumFromTraits<
+      WebMapServiceAvailableLayerDimensionsTraits
+    >[] = [];
+
+    if (!this.capabilities) {
+      return result;
+    }
+
+    const capabilitiesLayers = this.capabilitiesLayers;
+
+    for (const layerTuple of capabilitiesLayers) {
+      const layerName = layerTuple[0];
+      const layer = layerTuple[1];
+
+      const dimensions: ReadonlyArray<CapabilitiesDimension> = layer
+        ? this.capabilities.getInheritedValues(layer, "Dimension")
+        : [];
+
+      result.push({
+        layerName: layerName,
+        dimensions: dimensions
+          .filter(dim => dim.name !== "time")
+          .map(dim => {
+            return {
+              name: dim.name,
+              units: dim.units,
+              unitSymbol: dim.unitSymbol,
+              default: dim.default,
+              multipleValues: dim.multipleValues,
+              current: dim.current,
+              nearestValue: dim.nearestValue,
+              values: dim.text?.split(",")
+            };
+          })
+      });
+    }
+
+    return result;
+  }
+
+  @computed
   get availableStyles(): StratumFromTraits<
     WebMapServiceAvailableLayerStylesTraits
   >[] {
@@ -235,7 +268,9 @@ class GetCapabilitiesStratum extends LoadableStratum(
             ? undefined
             : createStratumInstance(LegendTraits, {
                 url: legendUri.toString(),
-                urlMimeType: legendMimeType
+                urlMimeType: legendMimeType,
+                title:
+                  (capabilitiesLayers.size > 1 && layer?.Title) || undefined // Add layer Title as legend title if showing multiple layers
               });
 
           return {
@@ -436,6 +471,7 @@ class GetCapabilitiesStratum extends LoadableStratum(
         layer,
         "Dimension"
       );
+
       const timeDimension = dimensions.find(
         dimension => dimension.name.toLowerCase() === "time"
       );
@@ -524,11 +560,6 @@ class DiffStratum extends LoadableStratum(WebMapServiceCatalogItemTraits) {
   }
 
   @computed
-  get disableStyleSelector() {
-    return this.catalogItem.isShowingDiff;
-  }
-
-  @computed
   get disableDateTimeSelector() {
     return this.catalogItem.isShowingDiff;
   }
@@ -546,7 +577,7 @@ class WebMapServiceCatalogItem
       )
     )
   )
-  implements Mappable {
+  implements Mappable, SelectableDimensions {
   /**
    * The collection of strings that indicate an Abstract property should be ignored.  If these strings occur anywhere
    * in the Abstract, the Abstract will not be used.  This makes it easy to filter out placeholder data like
@@ -626,44 +657,6 @@ class WebMapServiceCatalogItem
   }
 
   @computed
-  get styleSelector() {
-    if (this.availableStyles.length === 0) return undefined;
-    if (this.availableStyles[0].styles.length === 0) return undefined;
-
-    const userStrata: any = this.strata.get(CommonStrata.user);
-    let activeStyle = this.availableStyles[0].styles[0].name;
-    if (
-      isDefined(userStrata) &&
-      isDefined(userStrata.parameters) &&
-      isDefined(userStrata.parameters.styles)
-    ) {
-      activeStyle = userStrata.parameters.styles;
-    }
-    return {
-      name: "Styles",
-      id: `styles-${this.uniqueId}`,
-      activeStyleId: activeStyle,
-      availableStyles: this.availableStyles[0].styles.map(function(s) {
-        return {
-          name: s.title!,
-          id: s.name!
-        };
-      }),
-      chooseActiveStyle: (strata: string, newStyle: string) => {
-        let newParameters = {
-          styles: newStyle
-        };
-        if (isDefined(userStrata) && "parameters" in userStrata) {
-          newParameters = combine(newParameters, userStrata.parameters);
-        }
-        runInAction(() => {
-          this.setTrait(CommonStrata.user, "parameters", newParameters);
-        });
-      }
-    };
-  }
-
-  @computed
   get stylesArray(): ReadonlyArray<string> {
     if (Array.isArray(this.styles)) {
       return this.styles;
@@ -702,7 +695,9 @@ class WebMapServiceCatalogItem
   @computed
   get canDiffImages(): boolean {
     const hasValidDiffStyles = this.availableDiffStyles.some(diffStyle =>
-      this.styleSelector?.availableStyles.find(style => style.id === diffStyle)
+      this.styleSelectableDimensions?.[0]?.options.find(
+        style => style.id === diffStyle
+      )
     );
     return hasValidDiffStyles === true;
   }
@@ -855,18 +850,28 @@ class WebMapServiceCatalogItem
 
       console.log(`Creating new ImageryProvider for time ${time}`);
 
+      // Set dimensionParameters
+      const dimensionParameters = formatDimensionsForOws(this.dimensions);
+
+      if (time !== undefined) {
+        dimensionParameters.time = time;
+      }
+
       const diffModeParameters = this.isShowingDiff
         ? this.diffModeParameters
         : {};
-      const parameters: any = {
+
+      const parameters: { [key: string]: any } = {
         ...WebMapServiceCatalogItem.defaultParameters,
-        ...(this.parameters || {}),
-        ...diffModeParameters
+        ...this.parameters,
+        ...dimensionParameters
       };
 
-      if (time !== undefined) {
-        parameters.time = time;
+      if (isDefined(this.styles)) {
+        parameters.styles = this.styles;
       }
+
+      Object.assign(parameters, diffModeParameters);
 
       const maximumLevel = scaleDenominatorToLevel(this.minScaleDenominator);
 
@@ -909,6 +914,10 @@ class WebMapServiceCatalogItem
         url: proxyCatalogItemUrl(this, baseUrl.toString()),
         layers: this.layers || "",
         parameters: parameters,
+        getFeatureInfoParameters: {
+          ...dimensionParameters,
+          styles: this.styles
+        },
         tilingScheme: /*defined(this.tilingScheme) ? this.tilingScheme :*/ new WebMercatorTilingScheme(),
         maximumLevel: maximumLevel,
         rectangle: rectangle
@@ -963,6 +972,127 @@ class WebMapServiceCatalogItem
       return imageryProvider;
     }
   );
+
+  @computed
+  get styleSelectableDimensions(): SelectableDimension[] {
+    return this.availableStyles.map((layer, layerIndex) => {
+      let name = "Styles";
+
+      // If multiple layers -> prepend layer name to name
+      if (this.availableStyles.length > 1) {
+        // Attempt to get layer title from GetCapabilitiesStratum
+        const layerTitle =
+          layer.layerName &&
+          (this.strata.get(
+            GetCapabilitiesMixin.getCapabilitiesStratumName
+          ) as GetCapabilitiesStratum).capabilitiesLayers.get(layer.layerName)
+            ?.Title;
+
+        name = `${layerTitle ||
+          layer.layerName ||
+          `Layer ${layerIndex + 1}`} styles`;
+      }
+
+      return {
+        name,
+        id: `${this.uniqueId}-${layer.layerName}-styles`,
+        options: filterOutUndefined(
+          layer.styles.map(function(s) {
+            if (isDefined(s.name)) {
+              return {
+                name: s.title || s.name || "",
+                id: s.name as string
+              };
+            }
+          })
+        ),
+
+        // Set selectedId to value stored in `styles` trait for this `layerIndex` or the first available style value
+        // The `styles` parameter is CSV, a style for each layer
+        selectedId:
+          this.styles?.split(",")?.[layerIndex] || layer.styles[0].name,
+
+        setDimensionValue: (stratumId: string, newStyle: string) => {
+          runInAction(() => {
+            const styles = this.styleSelectableDimensions.map(
+              style => style.selectedId || ""
+            );
+            styles[layerIndex] = newStyle;
+            this.setTrait(stratumId, "styles", styles.join(","));
+          });
+        },
+        disable: this.isShowingDiff
+      };
+    });
+  }
+
+  @computed
+  get wmsDimensionSelectableDimensions(): SelectableDimension[] {
+    const dimensions: SelectableDimension[] = [];
+
+    // For each layer -> For each dimension
+    this.availableDimensions.forEach(layer => {
+      layer.dimensions.forEach(dim => {
+        // Only add dimensions if hasn't already been added (multiple layers may have the same dimension)
+        if (
+          !isDefined(dim.name) ||
+          dim.values.length < 2 ||
+          dimensions.findIndex(findDim => findDim.name === dim.name) !== -1
+        ) {
+          return;
+        }
+
+        dimensions.push({
+          name: dim.name,
+          id: `${this.uniqueId}-${dim.name}`,
+          options: dim.values.map(value => {
+            let name = value;
+            // Add units and unitSybol if defined
+            if (typeof dim.units === "string" && dim.units !== "") {
+              if (typeof dim.unitSymbol === "string" && dim.unitSymbol !== "") {
+                name = `${value} (${dim.units} ${dim.unitSymbol})`;
+              } else {
+                name = `${value} (${dim.units})`;
+              }
+            }
+            return {
+              name,
+              id: value
+            };
+          }),
+
+          // Set selectedId to value stored in `dimensions` trait, the default value, or the first available value
+          selectedId:
+            this.dimensions?.[dim.name]?.toString() ||
+            dim.default ||
+            dim.values[0],
+
+          setDimensionValue: (stratumId: string, newDimension: string) => {
+            let newDimensions: any = {};
+
+            newDimensions[dim.name!] = newDimension;
+
+            if (isDefined(this.dimensions)) {
+              newDimensions = combine(newDimensions, this.dimensions);
+            }
+            runInAction(() => {
+              this.setTrait(stratumId, "dimensions", newDimensions);
+            });
+          }
+        });
+      });
+    });
+
+    return dimensions;
+  }
+
+  @computed
+  get selectableDimensions() {
+    return filterOutUndefined([
+      ...this.styleSelectableDimensions,
+      ...this.wmsDimensionSelectableDimensions
+    ]);
+  }
 }
 
 function scaleDenominatorToLevel(
@@ -1102,6 +1232,32 @@ function formatMomentForWms(m: moment.Moment, duration: moment.Duration) {
   }
 
   return m.format();
+}
+
+/**
+ * Add `_dim` prefix to dimensions for OWS (WMS, WCS...) excluding time, styles and elevation
+ */
+export function formatDimensionsForOws(
+  dimensions: { [key: string]: string } | undefined
+) {
+  if (!isDefined(dimensions)) {
+    return {};
+  }
+  return Object.entries(dimensions).reduce<{ [key: string]: string }>(
+    (formattedDimensions, [key, value]) =>
+      // elevation is specified as simply "elevation", styles is specified as "styles"
+      // Other (custom) dimensions are prefixed with 'dim_'.
+      // See WMS 1.3.0 spec section C.3.2 and C.3.3.
+      {
+        formattedDimensions[
+          ["time", "styles", "elevation"].includes(key?.toLowerCase())
+            ? key
+            : `dim_${key}`
+        ] = value;
+        return formattedDimensions;
+      },
+    {}
+  );
 }
 
 function getServiceContactInformation(